+++ lib/objects/flying_attack_object.dart
@@ -1,40 +1,90 @@
+import 'dart:math';
+
 import 'package:bonfire/bonfire.dart';
 import 'package:bonfire/util/assets_loader.dart';
-import 'package:flame/extensions.dart';
 import 'package:flutter/widgets.dart';
 
-class FlyingAttackObject extends AnimatedObject with ObjectCollision, Lighting {
+class FlyingAttackObject extends GameComponent
+    with
+        WithSpriteAnimation,
+        WithAssetsLoader,
+        ObjectCollision,
+        Lighting,
+        Movement {
   final dynamic id;
   SpriteAnimation? flyAnimation;
-  Future<SpriteAnimation>? destroyAnimation;
-  final Direction direction;
+  Future<SpriteAnimation>? animationDestroy;
+  final Direction? direction;
   final double speed;
   final double damage;
   final AttackFromEnum attackFrom;
   final bool withDecorationCollision;
-  final VoidCallback? onDestroyedObject;
-  AssetsLoader? _loader = AssetsLoader();
-  final bool enableDiagonal;
-
-  final IntervalTick _timerVerifyCollision = IntervalTick(50);
+  final VoidCallback? onDestroy;
+  final bool enabledDiagonal;
+  final Vector2? destroySize;
+  double _cosAngle = 0;
+  double _senAngle = 0;
 
   FlyingAttackObject({
     required Vector2 position,
     required Vector2 size,
     required Future<SpriteAnimation> flyAnimation,
-    required this.direction,
+    this.direction,
+    double angle = 0,
     this.id,
-    this.destroyAnimation,
+    this.animationDestroy,
+    this.destroySize,
     this.speed = 150,
     this.damage = 1,
     this.attackFrom = AttackFromEnum.ENEMY,
     this.withDecorationCollision = true,
-    this.onDestroyedObject,
-    this.enableDiagonal = true,
+    this.onDestroy,
+    this.enabledDiagonal = true,
     LightingConfig? lightingConfig,
     CollisionConfig? collision,
   }) {
-    _loader?.add(AssetToLoad(flyAnimation, (value) {
+    loader?.add(AssetToLoad(flyAnimation, (value) {
+      return this.flyAnimation = value;
+    }));
+
+    this.position = position;
+    this.size = size;
+    this.angle = angle;
+    _cosAngle = cos(angle);
+    _senAngle = sin(angle);
+
+    if (lightingConfig != null) setupLighting(lightingConfig);
+
+    setupCollision(
+      collision ??
+          CollisionConfig(
+            collisions: [
+              CollisionArea.rectangle(
+                size: Vector2(width, height),
+              ),
+            ],
+          ),
+    );
+  }
+
+  FlyingAttackObject.byDirection({
+    required Vector2 position,
+    required Vector2 size,
+    required Future<SpriteAnimation> flyAnimation,
+    required this.direction,
+    this.id,
+    this.animationDestroy,
+    this.destroySize,
+    this.speed = 150,
+    this.damage = 1,
+    this.attackFrom = AttackFromEnum.ENEMY,
+    this.withDecorationCollision = true,
+    this.onDestroy,
+    this.enabledDiagonal = true,
+    LightingConfig? lightingConfig,
+    CollisionConfig? collision,
+  }) {
+    loader?.add(AssetToLoad(flyAnimation, (value) {
       return this.flyAnimation = value;
     }));
 
@@ -55,168 +105,95 @@ class FlyingAttackObject extends AnimatedObject with ObjectCollision, Lighting {
     );
   }
 
+  FlyingAttackObject.byAngle({
+    required Vector2 position,
+    required Vector2 size,
+    required Future<SpriteAnimation> flyAnimation,
+    required double angle,
+    this.id,
+    this.animationDestroy,
+    this.destroySize,
+    this.speed = 150,
+    this.damage = 1,
+    this.attackFrom = AttackFromEnum.ENEMY,
+    this.withDecorationCollision = true,
+    this.onDestroy,
+    this.enabledDiagonal = true,
+    LightingConfig? lightingConfig,
+    CollisionConfig? collision,
+  }) : direction = null {
+    loader?.add(AssetToLoad(flyAnimation, (value) {
+      return this.flyAnimation = value;
+    }));
+
+    this.position = position;
+    this.size = size;
+    this.angle = angle;
+    _cosAngle = cos(angle);
+    _senAngle = sin(angle);
+
+    if (lightingConfig != null) setupLighting(lightingConfig);
+
+    setupCollision(
+      collision ??
+          CollisionConfig(
+            collisions: [
+              CollisionArea.rectangle(
+                size: Vector2(width, height),
+              ),
+            ],
+          ),
+    );
+  }
+
   @override
   void update(double dt) {
     super.update(dt);
 
-    switch (direction) {
-      case Direction.left:
-        position = position.translate((speed * dt) * -1, 0);
-        break;
-      case Direction.right:
-        position = position.translate((speed * dt), 0);
-        break;
-      case Direction.up:
-        position = position.translate(0, (speed * dt) * -1);
-        break;
-      case Direction.down:
-        position = position.translate(0, (speed * dt));
-        break;
-      case Direction.upLeft:
-        if (enableDiagonal) {
-          position = position.translate((speed * dt) * -1, (speed * dt) * -1);
-        } else {
-          position = position.translate((speed * dt) * -1, 0);
-        }
-        break;
-      case Direction.upRight:
-        if (enableDiagonal) {
-          position = position.translate((speed * dt), (speed * dt) * -1);
-        } else {
-          position = position.translate((speed * dt), 0);
-        }
-
-        break;
-      case Direction.downLeft:
-        if (enableDiagonal) {
-          position = position.translate((speed * dt) * -1, (speed * dt));
-        } else {
-          position = position.translate((speed * dt) * -1, 0);
-        }
-
-        break;
-      case Direction.downRight:
-        if (enableDiagonal) {
-          position = position.translate((speed * dt), (speed * dt));
-        } else {
-          position = position.translate((speed * dt), 0);
-        }
-        break;
+    if (direction != null) {
+      moveFromDirection(direction!, enabledDiagonal: enabledDiagonal);
+    } else {
+      moveFromAngle(speed, angle);
     }
 
     if (!_verifyExistInWorld()) {
       removeFromParent();
-    } else {
-      _verifyCollision(dt);
     }
   }
 
-  void _verifyCollision(double dt) {
-    if (shouldRemove) return;
-    if (!_timerVerifyCollision.update(dt)) return;
-
-    bool destroy = false;
-
-    gameRef.visibleAttackables().where((a) {
-      final fromCorrect = (attackFrom == AttackFromEnum.ENEMY
-          ? a.receivesAttackFromEnemy()
-          : a.receivesAttackFromPlayer());
-
-      final overlap = a.rectAttackable().overlaps(rectCollision);
-
-      return fromCorrect && overlap;
-    }).forEach((enemy) {
-      enemy.receiveDamage(damage, id);
-      destroy = true;
-    });
-
-    if (withDecorationCollision && !destroy) {
-      destroy = isCollision().isNotEmpty;
-    }
-
-    if (destroy) {
-      if (destroyAnimation != null) {
-        Rect positionDestroy;
-        switch (direction) {
-          case Direction.left:
-            positionDestroy = Rect.fromLTWH(
-              rectCollision.left - (width / 2),
-              rectCollision.top - ((height - rectCollision.height) / 2),
-              width,
-              height,
-            );
-            break;
-          case Direction.right:
-            positionDestroy = Rect.fromLTWH(
-              rectCollision.right - (width / 2),
-              rectCollision.top - ((height - rectCollision.height) / 2),
-              width,
-              height,
-            );
-            break;
-          case Direction.up:
-            positionDestroy = Rect.fromLTWH(
-              rectCollision.left - ((width - rectCollision.width) / 2),
-              rectCollision.top - (height / 2),
-              width,
-              height,
-            );
-            break;
-          case Direction.down:
-            positionDestroy = Rect.fromLTWH(
-              rectCollision.left - ((width - rectCollision.width) / 2),
-              rectCollision.bottom + (height / 2),
-              width,
-              height,
-            );
-            break;
-          case Direction.upLeft:
-            positionDestroy = Rect.fromLTWH(
-              rectCollision.left - (width / 2),
-              rectCollision.top,
-              width,
-              height,
-            );
-            break;
-          case Direction.upRight:
-            positionDestroy = Rect.fromLTWH(
-              rectCollision.right - (width / 2),
-              rectCollision.top,
-              width,
-              height,
-            );
-            break;
-          case Direction.downLeft:
-            positionDestroy = Rect.fromLTWH(
-              rectCollision.left - (width / 2),
-              rectCollision.top,
-              width,
-              height,
-            );
-            break;
-          case Direction.downRight:
-            positionDestroy = Rect.fromLTWH(
-              rectCollision.right - (width / 2),
-              rectCollision.top,
-              width,
-              height,
-            );
-            break;
+  @override
+  bool onCollision(GameComponent component, bool active) {
+    if (component is Attackable && !component.shouldRemove) {
+      if (attackFrom == AttackFromEnum.ENEMY) {
+        if (component.receivesAttackFrom == ReceivesAttackFromEnum.ALL ||
+            component.receivesAttackFrom == ReceivesAttackFromEnum.ENEMY) {
+          component.receiveDamage(damage, id);
         }
-
-        gameRef.add(
-          AnimatedObjectOnce(
-            animation: destroyAnimation!,
-            position: positionDestroy.positionVector2,
-            size: positionDestroy.sizeVector2,
-            lightingConfig: lightingConfig,
-          ),
-        );
+      } else if (attackFrom == AttackFromEnum.PLAYER) {
+        if (component.receivesAttackFrom == ReceivesAttackFromEnum.ALL ||
+            component.receivesAttackFrom == ReceivesAttackFromEnum.PLAYER) {
+          component.receiveDamage(damage, id);
+        }
       }
-      setupCollision(CollisionConfig(collisions: []));
-      removeFromParent();
-      this.onDestroyedObject?.call();
+    } else if (!withDecorationCollision) {
+      return false;
     }
+    _destroyObject();
+    return true;
+  }
+
+  void _destroyObject() {
+    if (shouldRemove) return;
+    if (animationDestroy != null) {
+      if (direction != null) {
+        _destroyByDirection(direction!, dtUpdate);
+      } else {
+        _destroyByAngle();
+      }
+    }
+    setupCollision(CollisionConfig(collisions: []));
+    removeFromParent();
+    this.onDestroy?.call();
   }
 
   bool _verifyExistInWorld() {
@@ -242,9 +219,102 @@ class FlyingAttackObject extends AnimatedObject with ObjectCollision, Lighting {
 
   @override
   Future<void> onLoad() async {
-    await _loader?.load();
-    _loader = null;
+    await super.onLoad();
     animation = this.flyAnimation;
-    return super.onLoad();
+  }
+
+  void _destroyByDirection(Direction direction, double dt) {
+    Vector2 positionDestroy;
+
+    double biggerSide = max(width, height);
+    double addCenterX = 0;
+    double addCenterY = 0;
+
+    if (destroySize != null) {
+      addCenterX = ((size.x - destroySize!.x) / 2);
+      addCenterY = ((size.y - destroySize!.y) / 2);
+    }
+    switch (direction) {
+      case Direction.left:
+        positionDestroy = Vector2(
+          left - (biggerSide / 4) + addCenterX,
+          top + addCenterY,
+        );
+        break;
+      case Direction.right:
+        positionDestroy = Vector2(
+          left + (biggerSide / 4) + addCenterX,
+          top + addCenterY,
+        );
+        break;
+      case Direction.up:
+        positionDestroy = Vector2(
+          left + addCenterX,
+          top - (biggerSide / 4) + addCenterY,
+        );
+        break;
+      case Direction.down:
+        positionDestroy = Vector2(
+          left + addCenterX,
+          top + (biggerSide / 4) + addCenterY,
+        );
+        break;
+      case Direction.upLeft:
+        positionDestroy = Vector2(
+          left - (biggerSide / 4) + addCenterX,
+          top - (biggerSide / 4) + addCenterY,
+        );
+        break;
+      case Direction.upRight:
+        positionDestroy = Vector2(
+          left + (biggerSide / 4) + addCenterX,
+          top - (biggerSide / 4) + addCenterY,
+        );
+        break;
+      case Direction.downLeft:
+        positionDestroy = Vector2(
+          left - (biggerSide / 4) + addCenterX,
+          top + (biggerSide / 4) + addCenterY,
+        );
+        break;
+      case Direction.downRight:
+        positionDestroy = Vector2(
+          left + (biggerSide / 4) + addCenterX,
+          top + (biggerSide / 4) + addCenterY,
+        );
+
+        break;
+    }
+
+    gameRef.add(
+      AnimatedObjectOnce(
+        animation: animationDestroy!,
+        position: positionDestroy,
+        size: destroySize ?? size,
+        lightingConfig: lightingConfig,
+      ),
+    );
+  }
+
+  void _destroyByAngle() {
+    double nextX = (width / 4) * _cosAngle;
+    double nextY = (height / 4) * _senAngle;
+
+    Offset diffBase = Offset(
+          rectCollision.center.dx + nextX,
+          rectCollision.center.dy + nextY,
+        ) -
+        rectCollision.center;
+
+    final positionDestroy = position.translate(diffBase.dx, diffBase.dy);
+
+    gameRef.add(
+      AnimatedObjectOnce(
+        animation: animationDestroy!,
+        position: positionDestroy,
+        lightingConfig: lightingConfig,
+        size: destroySize ?? size,
+      ),
+    );
   }
 }