+++ lib/util/direction_animations/simple_direction_animation.dart
@@ -37,12 +37,11 @@ class SimpleDirectionAnimation {
   bool runToTheEndFastAnimation = false;
 
   double opacity = 1.0;
-
   bool _flipX = false;
   bool _flipY = false;
 
-  final bool enabledFlipX;
-  final bool enabledFlipY;
+  bool enabledFlipX;
+  bool enabledFlipY;
 
   SimpleDirectionAnimation({
     required FutureOr<SpriteAnimation> idleRight,
@@ -209,13 +208,13 @@ class SimpleDirectionAnimation {
       position: position,
       size: size,
       animation: animation,
-      flipX: flipX,
-      flipY: flipY,
       onFinish: () {
         onFinish?.call();
         _fastAnimation = null;
       },
     );
+    anim.isFlipVertical = flipY;
+    anim.isFlipHorizontal = flipX;
     await anim.onLoad();
     _fastAnimation = anim;
   }