+++ lib/util/extensions/game_component_extensions.dart
@@ -76,6 +76,7 @@ extension GameComponentExtensions on GameComponent {
     DirectionTextDamage direction = DirectionTextDamage.RANDOM,
     bool onlyUp = false,
   }) {
+    if (!hasGameRef) return;
     gameRef.add(
       TextDamageComponent(
         damage.toInt().toString(),
@@ -99,14 +100,18 @@ extension GameComponentExtensions on GameComponent {
 
   /// Execute the ranged attack using a component with animation
   void simpleAttackRangeByAngle({
-    required Future<SpriteAnimation> animationUp,
+    /// use animation facing right.
+    required Future<SpriteAnimation> animation,
     required Vector2 size,
-    required double radAngleDirection,
+
+    /// Use radians angle
+    required double angle,
     required double damage,
+    Vector2? destroySize,
     Future<SpriteAnimation>? animationDestroy,
     dynamic id,
     double speed = 150,
-    bool withCollision = true,
+    bool withDecorationCollision = true,
     VoidCallback? onDestroy,
     CollisionConfig? collision,
     LightingConfig? lightingConfig,
@@ -120,26 +125,27 @@ extension GameComponentExtensions on GameComponent {
 
     double displacement =
         max(initPosition.width / 2, initPosition.height / 2) + marginFromOrigin;
-    double nextX = displacement * cos(radAngleDirection);
-    double nextY = displacement * sin(radAngleDirection);
+    double nextX = displacement * cos(angle);
+    double nextY = displacement * sin(angle);
 
     Vector2 diffBase = Vector2(nextX, nextY);
 
     startPosition.add(diffBase);
     startPosition.add(Vector2(-size.x / 2, -size.y / 2));
-    gameRef.add(FlyingAttackAngleObject(
+    gameRef.add(FlyingAttackObject.byAngle(
       id: id,
       position: startPosition,
       size: size,
-      radAngle: radAngleDirection,
+      angle: angle,
       damage: damage,
       speed: speed,
       attackFrom: this is Player ? AttackFromEnum.PLAYER : AttackFromEnum.ENEMY,
       collision: collision,
-      withCollision: withCollision,
+      withDecorationCollision: withDecorationCollision,
       onDestroy: onDestroy,
-      flyAnimation: animationUp,
-      destroyAnimation: animationDestroy,
+      destroySize: destroySize,
+      flyAnimation: animation,
+      animationDestroy: animationDestroy,
       lightingConfig: lightingConfig,
     ));
   }
@@ -152,12 +158,13 @@ extension GameComponentExtensions on GameComponent {
     required Future<SpriteAnimation> animationDown,
     required Vector2 size,
     required Direction direction,
+    Vector2? destroySize,
     dynamic id,
     double speed = 150,
     double damage = 1,
     bool withCollision = true,
     bool enableDiagonal = true,
-    VoidCallback? destroy,
+    VoidCallback? onDestroy,
     CollisionConfig? collision,
     LightingConfig? lightingConfig,
     Future<SpriteAnimation>? animationDestroy,
@@ -231,19 +238,20 @@ extension GameComponentExtensions on GameComponent {
     }
 
     gameRef.add(
-      FlyingAttackObject(
+      FlyingAttackObject.byDirection(
         id: id,
         direction: attackDirection,
         flyAnimation: attackRangeAnimation,
-        destroyAnimation: animationDestroy,
+        animationDestroy: animationDestroy,
         position: startPosition,
         size: size,
         damage: damage,
         speed: speed,
-        enableDiagonal: enableDiagonal,
+        enabledDiagonal: enableDiagonal,
         attackFrom:
             this is Player ? AttackFromEnum.PLAYER : AttackFromEnum.ENEMY,
-        onDestroyedObject: destroy,
+        onDestroy: onDestroy,
+        destroySize: destroySize,
         withDecorationCollision: withCollision,
         collision: collision,
         lightingConfig: lightingConfig,